Split soc and board modules
