+++ runners/embedded/Cargo.toml
@@ -94,11 +94,10 @@ format-filesystem = []
 
 alloc = ["alloc-cortex-m"]
 
-board-nk3am = ["soc-nrf52840"]
-
+board-nk3am = ["soc-nrf52"]
 board-nk3xn = ["soc-lpc55"]
 
-soc-nrf52840 = ["nrf52840-hal", "nrf52840-pac", "rand_chacha", "se05x?/nrf"]
+soc-nrf52 = ["nrf52840-hal", "nrf52840-pac", "rand_chacha", "se05x?/nrf"]
 soc-lpc55 = ["lpc55-hal", "lpc55-pac", "fm11nc08", "systick-monotonic", "se05x?/lpc55", "rand_chacha"]
 
 se050 = ["dep:se05x", "apps/se050"]
@@ -124,7 +123,7 @@ log-semihosting = ["cortex-m-semihosting"]
 [[bin]]
 name = "nrf52_runner"
 path = "src/bin/app-nrf.rs"
-required-features = ["soc-nrf52840"]
+required-features = ["soc-nrf52"]
 
 [[bin]]
 name = "lpc55_runner"
